Convert streamer from module into package
This is a long-overdue conversion of the unwieldy `h.streamer` module
into a package.
